requirements.txt
python-dotenv
streamlit
Pillow
openrouteservice
requests
groq
beautifulsoup4